Rename pmemkv.cc and pmemkv.h to libpmemkv.cc and libpmemkv.h
pmemkv.h is already installed as libpmemkv.h
