+++ src/engines/blackhole.h
@@ -32,7 +32,7 @@
 
 #pragma once
 
-#include "../pmemkv.h"
+#include "../libpmemkv.h"
 
 namespace pmemkv {
 namespace blackhole {
