+++ src/libpmemkv.h
